# apps/api/docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ownpace — managed API
  version: 0.1.0-rc.1
  summary: Migrate mail, calendar, contacts and files off US cloud, on your own schedule.
  description: |
    The multi-tenant managed edition. Every request is scoped to one tenant by
    the bearer token; the database enforces the same boundary again with
    row-level security, so a token for tenant A cannot read tenant B even if a
    route forgets to filter.

    **Two ideas run through the whole surface and explain most of its shape:**

    - *Nothing consequential happens without a person.* Deleting on the target,
      finishing a migration and moving billing money are gated behind explicit
      evidence, an owner/admin role, or both. Endpoints that could destroy data
      refuse by default and say what is missing.
    - *A failure is reported as itself.* Routes return the reason verbatim
      rather than an empty success. "I found nothing" and "I could not look"
      are different answers and never share a status code.

    Long-running work (sync, cutover, verification, applying a deletion) is
    enqueued, not performed in the request: those endpoints answer `202` with
    an identifier, and a matching read endpoint polls the outcome.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Ownpace
    url: https://github.com/Robbes/Ownpace
servers:
  - url: https://{host}
...
    variables:
      host:
        default: localhost:3001
...
security:
  - bearerAuth: []
...
paths:
...
  /api/billing/usage:
    get:
      tags: [Billing]
      summary: This period's usage and cost
      description: |
        Priced at **this tenant's agreed rates** — the snapshot pinned when the
        tenant was first billed, not the operator's current price list. An
        operator raising prices for new customers does not re-price an existing
        one.
      responses:
        '200':
          description: Usage and the cost it implies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  usage: { $ref: '#/components/schemas/Usage' }
                  currentCost: { $ref: '#/components/schemas/Cost' }
                  period: { type: string, examples: ['2026-08'] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/billing/usage/history:
    get:
      tags: [Billing]
      summary: Usage by period
      responses:
        '200':
          description: One entry per period, newest first.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/billing/estimate:
    post:
      tags: [Billing]
      summary: What projected usage would cost
      description: Estimated at this tenant's agreed rates, so the number matches the invoice they will get.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storageUsedGB: { type: number }
                egressGB: { type: number }
                computeHours: { type: number }
                syncCount: { type: number }
      responses:
        '200':
          description: The estimate, in integer cents.
          content:
            application/json:
              schema:
                type: object
                properties:
                  estimate: { type: integer }
                  breakdown: { $ref: '#/components/schemas/Cost' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/billing/invoices:
    get:
      tags: [Billing]
      summary: List invoices
      responses:
        '200':
          description: The tenant's invoices.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/billing/invoices/generate:
    post:
      tags: [Billing]
      summary: Generate or refresh a period's invoice
      description: |
        Owner/admin. An invoice already paid or void is left untouched rather
        than regenerated — a settled document does not move because usage was
        recounted.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                period: { type: string, pattern: '^\d{4}-\d{2}$', examples: ['2026-08'] }
      responses:
        '201':
          description: The generated invoice.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/billing/invoices/{invoiceId}:
    parameters:
      - name: invoiceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Billing]
      summary: Get one invoice
      responses:
        '200':
          description: The invoice.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/billing/invoices/{invoiceId}/pay:
    parameters:
      - name: invoiceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Billing]
      summary: Start a payment
      description: Owner/admin. Returns the provider's checkout URL to send the payer to.
      responses:
        '200':
          description: The payment, with a checkout URL.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/billing/payment-methods:
    get:
      tags: [Billing]
      summary: List payment methods
      responses:
        '200':
          description: The tenant's payment methods.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Billing]
      summary: Add a payment method
      description: Owner/admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '201':
          description: The stored method.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/billing/payment-methods/{paymentMethodId}/default:
    parameters:
      - name: paymentMethodId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Billing]
      summary: Make this the default payment method
      description: Owner/admin.
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A tenant-scoped JWT. Local deployments sign it with `JWT_SECRET`;
        managed SSO validates against a JWKS instead.

        The token's subject must be a MEMBER of the tenant it names — an
        arbitrary subject with a valid signature is refused. Roles (`owner`,
        `admin`, `member`, `viewer`) gate the routes marked owner/admin above,
        for reads as well as writes on billing.
...
  responses:
    BadRequest:
      description: The request was refused, with the reason.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid, or non-member token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but this role may not do this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource in this tenant.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      description: Every refusal has this shape. `message` is meant to be shown to a person as-is.
      properties:
        error: { type: string, examples: ['Conflict'] }
        message:
          type: string
          examples: ['Mapping is paused — review the discovery counts and start it first (POST /start).']
      required: [error, message]
...
    Usage:
      type: object
      properties:
        tenantId: { type: string, format: uuid }
        period: { type: string, examples: ['2026-08'] }
        storageUsedGB: { type: number }
        egressGB: { type: number }
        computeHours:
          type: number
          description: Time the passes actually took, summed from their own start and finish stamps.
        syncCount: { type: integer }
        lastUpdated: { type: string, format: date-time }
    Cost:
      type: object
      description: Integer cents throughout — no float reaches an invoice.
      properties:
        baseFee: { type: integer }
        storage: { type: integer }
        egress: { type: integer }
        compute: { type: integer }
        subtotal: { type: integer }
        taxRate: { type: number, examples: [0.21] }
        tax: { type: integer }
        total: { type: integer }
      required: [baseFee, subtotal, taxRate, tax, total]
...